+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix missing GraphQL type schema [#376](https://github.com/hypermodeAI/runtime/pull/376)
 - Add FunctionExecutionDurationMillisecondsSummary metric [#377](https://github.com/hypermodeAI/runtime/pull/377)
+- Fix field alignment issue [#378](https://github.com/hypermodeAI/runtime/pull/378)
 
 ## 2024-09-18 - Version 0.12.1